+++ app/src/main/java/io/horizontalsystems/bankwallet/core/Caution.kt
@@ -1,0 +1,7 @@
+package io.horizontalsystems.bankwallet.core
+
+data class Caution(val text: String, val type: Type) {
+    enum class Type {
+        Error, Warning
+    }
+}
