+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/eip20approve/Eip20ApproveFragment.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.swap.approve
+package io.horizontalsystems.bankwallet.modules.eip20approve
 
 import android.os.Parcelable
 import androidx.compose.foundation.layout.Column
