+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/action/ActionRevoke.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.slideFromBottomForResult
 import io.horizontalsystems.bankwallet.entities.CoinValue
-import io.horizontalsystems.bankwallet.modules.swap.approve.Eip20RevokeConfirmFragment
+import io.horizontalsystems.bankwallet.modules.eip20revoke.Eip20RevokeConfirmFragment
 import io.horizontalsystems.marketkit.models.Token
 import java.math.BigDecimal
