+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/settings/SwapSlippageViewModel.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.swap.settings
+package io.horizontalsystems.bankwallet.modules.multiswap.settings
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -6,9 +6,10 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.Caution
 import io.horizontalsystems.bankwallet.core.providers.Translator
 import io.horizontalsystems.bankwallet.entities.DataState
-import io.horizontalsystems.bankwallet.modules.swap.settings.ui.InputButton
+import io.horizontalsystems.bankwallet.modules.multiswap.settings.ui.InputButton
 import io.reactivex.Observable
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.rx2.asFlow
